requirements_rsql_app.txt
# RSQL Lineage Converter - Gradio Web App
gradio>=4.0.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
pydantic>=2.0
python-dotenv>=1.0.0
openai>=1.0.0
networkx>=3.0
matplotlib>=3.5.0
